+++ src/dataloaders/stock-portfolio.dataloader.ts
@@ -1,0 +1,12 @@
+import { IStockPortfolio } from "@/mongodb";
+import DataLoader from "dataloader";
+import { batchById, batchByIdAndFilters, IObjectFilterKey } from "./utils";
+
+export const stockPortfolioById = (): DataLoader<string, IStockPortfolio & IDataNode> => {
+	return new DataLoader((keys: string[]) => batchById("StockPortfolio", keys));
+};
+
+export const stockPortfolio = (): DataLoader<
+	IObjectFilterKey,
+	(IStockPortfolio & IDataNode) | null
+> => new DataLoader((keys) => batchByIdAndFilters("StockPortfolio", keys));