+++ src/server/index.ts
@@ -8,6 +8,8 @@ import { applyMiddlewares } from "./middlewares";
 import { applyRoutes } from "./routes";
 
 const DEFAULT_PORT: number = 8080;
+const MAX_GRAPHQL_COMPLEXITY: number = 500;
+const MAX_GRAPHQL_DEPTH: number = 10;
 
 const applyGraphQL = (server: Server): void => {
 	const apolloServer: ApolloServer = getApolloServer<
@@ -15,7 +17,9 @@ const applyGraphQL = (server: Server): void => {
 		{ ctx: ParameterizedContext<IWithUser> }
 	>(ApolloServer, {
 		getHeaders: ({ ctx: { header } }) => header,
-		getKoaCtx: ({ ctx }) => ctx
+		getKoaCtx: ({ ctx }) => ctx,
+		maxComplexity: MAX_GRAPHQL_COMPLEXITY,
+		maxDepth: MAX_GRAPHQL_DEPTH
 	});
 
 	apolloServer.applyMiddleware({ app: server.app });