+++ CHANGELOG.md
@@ -5,3 +5,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## 0.0.1 - 2019-10-28
+### Added
+- Added authentication mutations.
+- Added Create and Read operations for stock portfolios.
+- Added query cost analysis.
+- Added query rate limiting.
+- Added query depth limit.
+- Added query permissions.
+
+### Fixed
+- Fixed resources not stopping on completed request.
+
+[Unreleased]: https://github.com/mua-inc/app_server/compare/v0.0.1...HEAD